feature/Endpoints/Multilingual
Added new endpoints for tourists. Implemented multilingualism on the main page
