+++ client/package.json
@@ -14,9 +14,11 @@
     "axios": "^1.8.3",
     "bcrypt": "^5.1.1",
     "helmet": "^8.1.0",
+    "i18next": "^24.2.3",
     "jsonwebtoken": "^9.0.2",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-i18next": "^15.4.1",
     "react-router-dom": "^7.3.0"
   },
   "devDependencies": {
