+++ client/src/components/Button.jsx
@@ -1,6 +1,8 @@
 
-const Button = ({ children }) => {
-    return <button className="w3-bar-item w3-red">{children}</button>;
+const Button = ({ children, onClick  }) => {
+    return <button className="w3-bar-item w3-red"onClick={onClick}>
+      {children}
+    </button>
   };
-// <Button>TEXT</Button>
+
 export default Button;
