+++ client/src/locales/en.json
@@ -1,0 +1,4 @@
+{
+    "select_date_city": "Select date and city for your tour",
+    "become_guide": "Become a Guide"
+}
