+++ client/src/locales/fi.json
@@ -1,0 +1,5 @@
+{
+    "select_date_city": "Valitse päivämäärä ja kaupunki",
+    "become_guide": "Ryhdy oppaaksi"
+}
+
