+++ client/src/locales/ru.json
@@ -1,0 +1,4 @@
+{
+"select_date_city": "Выберите дату и город для вашего тура",
+"become_guide": "Стань гидом"
+}
