+++ client/src/main.jsx
@@ -2,7 +2,9 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from "react-router-dom";
 import App from './App.jsx';
+import './i18n'
 import './index.css'
+
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
